requirements.txt
# LLM
groq>=0.9.0

# Vector DB
chromadb>=0.5.0

# Embeddings
# torch + sentence-transformers + numpy are installed separately
# in Dockerfile with pinned versions — do NOT add them here
sentence-transformers>=3.0.0

# NumPy pinned to <2 for torch compatibility
numpy<2

# Web scraping + URL summarization
beautifulsoup4>=4.12.0
requests>=2.32.0

# File summarization
pypdf>=4.0.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0

# Web search
ddgs>=6.0.0

# Config
python-dotenv>=1.0.0

# Google Calendar + Gmail
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Date parsing
python-dateutil>=2.9.0
dateparser>=1.2.0

# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# MCP
mcp[cli]>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.0.0